+++ parser.h
@@ -17,7 +17,7 @@ class CParser
 	char *ConvertStr(char *str);
 	double ConvertValue(int signal_id,double data);
 	void SetValidData();
-	void SetFunction(int id_function, float *values);
+	void SetFunction(int id_function, double *values);
 	void Reset(float *tab);
 	double CParser::DD( float DM );
